Change from __dirname to process.cwd()
We're using swagger-node from Docker and are using a local copy of the swagger script from `./node_modules/.bin/swagger` to run `./node_modules/.bin/swagger project edit`. This doesn't serve `swagger-editor` correctly as `__dirname` resolves to `./node_modules/swagger` as this is where the script is running. We've made the above change to use `process.cwd()` so that we can reference swagger-editor correctly.